pyproject.toml
[project]
name = "mcp-cca"
version = "0.1.0"
description = "Add your description here"
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "code-context-analyzer==0.3.3",
    "isort>=6.0.1",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
]
